Add details views on profiles and policies
