+++ frontend/src/pages/cert-manager/CertAuthDetailsByIDPage/route.tsx
@@ -6,7 +6,9 @@ import { CertAuthDetailsByIDPage } from "./CertAuthDetailsByIDPage";
 
 const caDetailsSearchSchema = z.object({
   from: z.enum(["settings", "profile"]).optional(),
-  profileId: z.string().optional()
+  profileId: z.string().optional(),
+  profileFrom: z.enum(["settings", "application"]).optional(),
+  profileApplicationName: z.string().optional()
 });
 
 export const Route = createFileRoute(
