+++ frontend/src/pages/cert-manager/CertificateProfileDetailsByIDPage/components/index.tsx
@@ -1,4 +1,2 @@
 export { ProfileDefaultsSection } from "./ProfileDefaultsSection";
-export { ProfileDetailsSection } from "./ProfileDetailsSection";
-export { ProfileIssuerSection } from "./ProfileIssuerSection";
-export { ProfilePolicySection } from "./ProfilePolicySection";
+export { ProfileOverviewSection } from "./ProfileOverviewSection";
